params: note about json tag for eip160(F)Block
This is a bit of a sticky situation. Solves are either:
- use without 'F' here, break pattern, stay backwards
compatible for existing multi-geth clients, but incompatible
with ETH if they accept proposed changes as-is
- use with 'F' here, break backwards compatibility with
existing multi-geth clients, but be compatible with ETH as proposed
- push change without 'F' to ETH for review, keeping
in mind that it makes no sense for them to tangle their
code b/c of an establish multi-geth pattern
- push change without 'F' for any of the EIP160,161,170 to
eth. still breaks pattern for them

The root cause of this is that EIP158 namespace was used
to symbolize 3 separate EIPs... none of which ended up being
named 'EIP158'. Yea.